feat: ensure typescript support
and migrate some files to .tsx
