+++ .eslintrc.js
@@ -1,5 +1,5 @@
 module.exports = {
-  parser: '@babel/eslint-parser',
+  parser: '@typescript-eslint/parser',
   parserOptions: {
     requireConfigFile: false,
     babelOptions: {
@@ -7,7 +7,7 @@ module.exports = {
     }
   },
   extends: ['react-app', 'standard', 'plugin:jsx-a11y/recommended'],
-  plugins: ['jsx-a11y', 'storybook'],
+  plugins: ['@typescript-eslint', 'jsx-a11y', 'storybook'],
   // ignore .ts files because it fails to parse it.
   ignorePatterns: 'src/**/*.ts',
   rules: {
