+++ .storybook/main.js
@@ -37,6 +37,18 @@ const storybookConfig = {
     postcss: false,
     storyStoreV7: true
   },
+  /**
+   * @see https://storybook.js.org/docs/react/configure/typescript
+   */
+  typescript: {
+    check: true,
+    checkOptions: {},
+    reactDocgen: 'react-docgen-typescript',
+    reactDocgenTypescriptOptions: {
+      shouldExtractLiteralValuesFromEnum: true,
+      propFilter: (prop) => (prop.parent ? !/node_modules/.test(prop.parent.fileName) : true),
+    },
+  },
   webpackFinal: async (config) => {
     return webpackOverride({
       ...config,
