+++ @types/redux-bundler-react/index.d.ts
@@ -1,0 +1,11 @@
+declare module 'redux-bundler-react' {
+  interface Connect {
+  }
+
+  interface Provider extends React.Provider {
+  }
+  
+  declare export const Provider: Provider;
+
+  declare export const connect: Connect
+}
