+++ package.json
@@ -134,10 +134,14 @@
     "@storybook/store": "^6.5.9",
     "@storybook/test-runner": "^0.7.0",
     "@svgr/cli": "^5.4.0",
-    "@types/node": "^14.0.27",
+    "@types/jest": "^29.0.3",
+    "@types/node": "^14.18.31",
     "@types/path-browserify": "^1.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.30.7",
-    "@typescript-eslint/parser": "^5.30.7",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-helmet": "^6.1.5",
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
     "assert": "^2.0.0",
     "autoprefixer": "^10.4.7",
     "babel-eslint": "^10.1.0",
@@ -181,7 +185,7 @@
     "semantic-release": "^19.0.3",
     "shx": "^0.3.3",
     "stream-browserify": "^3.0.0",
-    "typescript": "^4.1.3",
+    "typescript": "^4.8.4",
     "wait-on": "^6.0.1",
     "webpack": "^5.73.0",
     "webpack-bundle-analyzer": "^3.8.0",
