--- src/components/box/Box.js
@@ -1,19 +1,0 @@
-import React from 'react'
-import ErrorBoundary from '../error/ErrorBoundary'
-
-export const Box = ({
-  className = 'pa4',
-  style,
-  children,
-  ...props
-}) => {
-  return (
-    <section className={className} style={{ background: '#fbfbfb', ...style }}>
-      <ErrorBoundary>
-        {children}
-      </ErrorBoundary>
-    </section>
-  )
-}
-
-export default Box
