--- src/icons/GlyphAdd.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const GlyphAdd = props => (
-  <svg viewBox='0 0 100 100' {...props}>
-    <path d='M71.13 28.87a29.88 29.88 0 1 0 0 42.26 29.86 29.86 0 0 0 0-42.26zm-18.39 37.6h-5.48V52.74H33.53v-5.48h13.73V33.53h5.48v13.73h13.73v5.48H52.74z' />
-  </svg>
-)
-
-export default GlyphAdd
